requirements.txt
lxml==4.6.4
python-docx==0.8.11
tk==0.1.0
